Updated surveys folder
