+++ packages/hosting/src/router/index.js
@@ -4,6 +4,7 @@ import { AdminLayout } from "../components/layout/AdminLayout";
 import { PrivateRoute } from "./PrivateRoute";
 import * as A from "../pages";
 import { Page404 } from "../pages/404";
+import { OrganizationalClimateStudyIntegration } from "../pages";
 
 export const Router = () => {
   return (
@@ -275,6 +276,15 @@ export const Router = () => {
         element={
           <AdminLayout>
             <A.OrganizationalClimateStudiesIntegration />
+          </AdminLayout>
+        }
+      />
+      <Route
+        exact
+        path="/surveys/organizational-climate-studies/:organizationalClimateStudyId"
+        element={
+          <AdminLayout>
+            <A.OrganizationalClimateStudyIntegration />
           </AdminLayout>
         }
       />
